fix: isolate blogSlugHash from remark imports
Move blogSlugHash into a remark-free module so blog-loader and layout
routes no longer pull unified/remark-parse at module init.
